+++ index.php
@@ -20,7 +20,7 @@
     require "route.php";
     require "locale.php";
 
-    $emoncms_version = ($feed_settings['redisbuffer']['enabled'] ? "low-write " : "") . "9 RC | 2015.09.07";
+    $emoncms_version = ($feed_settings['redisbuffer']['enabled'] ? "low-write " : "") . "9 RC2 | 2015.09.09";
 
     $path = get_application_path();
     require "Lib/EmonLogger.php";
